# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Example API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Список всех пользователей",
                "operationId": "getUsersList",
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/userData"
                                            }
                                        },
                                        "error": {
                                            "type": "string",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/500Error"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Создать пользователя",
                "operationId": "createUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/storeUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Пользователь создан",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "$ref": "#/components/schemas/userData"
                                        },
                                        "error": {
                                            "type": "string",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/storeUserValidationErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/500Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Получить пользователя по ID",
                "operationId": "getUserById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID пользователя",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "$ref": "#/components/schemas/userData"
                                        },
                                        "error": {
                                            "type": "string",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/404Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/500Error"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Обновить пользователя",
                "operationId": "updateUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID пользователя",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/updateUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Пользователь обновлен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "$ref": "#/components/schemas/userData"
                                        },
                                        "error": {
                                            "type": "string",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/updateUserValidationErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/404Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/500Error"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Удалить пользователя",
                "operationId": "deleteUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID пользователя",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Удалено без содержимого"
                    },
                    "404": {
                        "description": "Не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/404Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/500Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/users/{id}/posts": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Посты пользователя",
                "operationId": "getUserPosts",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID пользователя",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/postData"
                                            }
                                        },
                                        "error": {
                                            "type": "string",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/404Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/500Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/users/{id}/comments": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Комментарии пользователя",
                "operationId": "getUserComments",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID пользователя",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/commentData"
                                            }
                                        },
                                        "error": {
                                            "type": "string",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/404Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/500Error"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "500Error": {
                "properties": {
                    "result": {
                        "type": "array",
                        "items": {},
                        "example": []
                    },
                    "error": {
                        "type": "string",
                        "example": "Error string"
                    }
                },
                "type": "object"
            },
            "404Error": {
                "properties": {
                    "result": {
                        "type": "array",
                        "items": {},
                        "example": []
                    },
                    "error": {
                        "type": "string",
                        "example": "Not found"
                    }
                },
                "type": "object"
            },
...
            "storeUserRequest": {
                "description": "Запрос на создание пользователя",
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "description": "Имя пользователя",
                        "type": "string",
                        "example": "Alice"
                    }
                },
                "type": "object"
            },
            "storeUserValidationErrorResponse": {
                "description": "Ответ при ошибках валидации при создании пользователя",
                "properties": {
                    "result": {
                        "type": "array",
                        "items": {},
                        "example": []
                    },
                    "error": {
                        "properties": {
                            "name": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "The name field is required."
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "updateUserRequest": {
                "description": "Запрос на обновление пользователя",
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "description": "Новое имя пользователя",
                        "type": "string",
                        "example": "Bob"
                    }
                },
                "type": "object"
            },
            "updateUserValidationErrorResponse": {
                "description": "Ответ при ошибках валидации при обновлении пользователя",
                "properties": {
                    "result": {
                        "type": "array",
                        "items": {},
                        "example": []
                    },
                    "error": {
                        "properties": {
                            "name": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "The name field is required."
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "commentData": {
                "description": "Данные комментария",
                "required": [
                    "id",
                    "post_id",
                    "user",
                    "body",
                    "created"
                ],
                "properties": {
                    "id": {
                        "description": "Идентификатор комментария",
                        "type": "integer",
                        "example": 42
                    },
                    "post_id": {
                        "description": "Идентификатор поста",
                        "type": "integer",
                        "example": 10
                    },
                    "user": {
                        "$ref": "#/components/schemas/userData"
                    },
                    "body": {
                        "description": "Текст комментария",
                        "type": "string",
                        "example": "Отличный пост!"
                    },
                    "created": {
                        "description": "Дата и время создания",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-05T11:15:00Z"
                    }
                },
                "type": "object"
            },
            "postData": {
                "description": "Данные поста",
                "required": [
                    "id",
                    "user",
                    "body",
                    "created"
                ],
                "properties": {
                    "id": {
                        "description": "Идентификатор поста",
                        "type": "integer",
                        "example": 10
                    },
                    "user": {
                        "$ref": "#/components/schemas/userData"
                    },
                    "body": {
                        "description": "Содержимое поста",
                        "type": "string",
                        "example": "Hello, world!"
                    },
                    "created": {
                        "description": "Дата и время создания",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-05T10:00:00Z"
                    }
                },
                "type": "object"
            },
            "userData": {
                "description": "Данные пользователя",
                "required": [
                    "id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "description": "Идентификатор пользователя",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Имя пользователя",
                        "type": "string",
                        "example": "Alice"
                    }
                },
                "type": "object"
            }
        }
    },
...
}